@@ -9,7 +9,10 @@
 package imapserver
 
 import (
+	"log"
+
 	idle "github.com/emersion/go-imap-idle"
+	move "github.com/emersion/go-imap-move"
 	"github.com/emersion/go-imap/server"
 	"github.com/emersion/go-sasl"
 )
@@ -17,6 +20,7 @@ import (
 type IMAPServer struct {
 	server  *server.Server
 	backend *Backend
+	notify  *IMAPNotify
 }
 
 func NewIMAPServer(backend *Backend, addr string, insecure bool) (*IMAPServer, *IMAPNotify, error) {
@@ -24,12 +28,13 @@ func NewIMAPServer(backend *Backend, addr string, insecure bool) (*IMAPServer, *
 		server:  server.New(backend),
 		backend: backend,
 	}
-	notify := NewIMAPNotify(s.server, backend.Log)
+	s.notify = NewIMAPNotify(s.server, backend.Log)
 	s.server.Addr = addr
 	s.server.AllowInsecureAuth = insecure
 	//s.server.Debug = os.Stdout
 	s.server.Enable(idle.NewExtension())
-	s.server.Enable(notify)
+	s.server.Enable(move.NewExtension())
+	// s.server.Enable(s.notify)
 	s.server.EnableAuth(sasl.Login, func(conn server.Conn) sasl.Server {
 		return sasl.NewLoginServer(func(username, password string) error {
 			_, err := s.backend.Login(nil, username, password)
@@ -38,10 +43,10 @@ func NewIMAPServer(backend *Backend, addr string, insecure bool) (*IMAPServer, *
 	})
 	go func() {
 		if err := s.server.ListenAndServe(); err != nil {
-			backend.Log.Printf("IMAP server stopped: %v", err)
+			log.Fatal(err)
 		}
 	}()
-	return s, notify, nil
+	return s, s.notify, nil
 }
 
 // Close closes the IMAP server
