@@ -9,7 +9,9 @@
 package smtpsender
 
 import (
+	"database/sql"
 	"encoding/hex"
+	"errors"
 	"fmt"
 	"log"
 	"net/mail"
@@ -145,12 +147,14 @@ func (q *Queue) run() {
 		return
 	}
 
-	q.queues.Log.Println("There are", len(refs), "mail(s) queued for", q.destination)
-
 	for _, ref := range refs {
 		_, mail, err := q.queues.Storage.MailSelect("Outbox", ref.ID)
 		if err != nil {
-			q.queues.Log.Println("Failed to get mail", ref.ID, "due to error:", err)
+			if errors.Is(err, sql.ErrNoRows) {
+				q.queues.Storage.QueueDeleteDestinationForID("Outbox", ref.ID)
+			} else {
+				q.queues.Log.Println("Failed to get mail", ref.ID, "due to error:", err)
+			}
 			continue
 		}
 
